Guard can now catch player if he's a shoplifter.

+++ MyGraphics/Source/SP2.h
@@ -363,7 +363,7 @@ private:
 	float remaindingAmount;
 	float amountOvershot;
 	//Villain
-	int objectsDestroyed;
+	unsigned long long objectsDestroyed;
 	bool isCaught;
 	//Guard
 	bool villainCaught;
